refactor(javascript): consolidate parser semantic helpers
