+++ crates/rspack_plugin_javascript/src/visitors/dependency/parser/mod.rs
@@ -4,7 +4,6 @@ pub mod ast;
 mod call_hooks_name;
 pub mod estree;
 mod location_advancer;
-mod semantic;
 mod walk;
 mod walk_block_pre;
 mod walk_module_pre;
@@ -1047,8 +1046,49 @@ impl<'parser> JavascriptParser<'parser> {
     })
   }
 
+  fn get_name_info_from_root(&mut self, root: ExprRef) -> Option<NameInfo<'_>> {
+    let ExprRef::Ident(identifier) = root else {
+      return self.get_name_info_from_variable(root.get_root_name(self.ast.ast)?);
+    };
+    let Some(state) = self.definitions_db.resolve_identifier(self.ast, identifier) else {
+      let ast = self.ast.ast;
+      return Some(NameInfo {
+        name: ast.get_utf8(identifier.name(ast)),
+        info: None,
+      });
+    };
+    let info = self.definitions_db.expect_get_variable(state);
+    if !info.is_free() && !info.is_tagged() {
+      return None;
+    }
+    Some(NameInfo {
+      name: info.name?,
+      info: Some(info),
+    })
+  }
+
   pub fn define_variable(&mut self, name: Atom) {
     self.definitions_db.define(name);
+  }
+
+  fn define_variable_identifier(&mut self, identifier: BindingIdentifier) {
+    self.definitions_db.define_identifier(self.ast, identifier);
+  }
+
+  fn pre_define_variable_identifier(&mut self, identifier: BindingIdentifier) {
+    self
+      .definitions_db
+      .pre_define_identifier(self.ast, identifier);
+  }
+
+  fn activate_semantic_scope_bindings(&mut self) {
+    self.definitions_db.activate_scope_bindings(self.ast);
+  }
+
+  fn define_function_declaration(&mut self, identifier: BindingIdentifier) {
+    self
+      .definitions_db
+      .define_function_declaration(self.ast, identifier);
   }
 
   pub fn set_variable(&mut self, name: Atom, variable: ExportedVariableInfo) {
